Created models subpackage
