+++ camel/models/stub_model.py
@@ -1,0 +1,42 @@
+# =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
+# Licensed under the Apache License, Version 2.0 (the “License”);
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an “AS IS” BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
+from typing import Any, Dict, List
+
+from .base_model import BaseModel
+
+
+class StubModel(BaseModel):
+    r"""A dummy model used for unit tests."""
+
+    def __init__(self, *args, **kwargs) -> None:
+        r"""All arguments are unused for the dummy model."""
+        super().__init__()
+
+    def run(self, messages: List[Dict]) -> Dict[str, Any]:
+        r"""Run fake inference by returning a fixed string.
+        All arguments are unused for the dummy model.
+
+        Returns:
+            Dict[str, Any]: Response in the OpenAI API format.
+        """
+        ARBITRARY_STRING = "Lorem Ipsum"
+
+        return dict(
+            id="stub_model_id",
+            usage=dict(),
+            choices=[
+                dict(finish_reason="stop",
+                     message=dict(content=ARBITRARY_STRING, role="assistant"))
+            ],
+        )
